digit-recognition/requirements.txt
scikit-learn
numpy
joblib
Pillow
